% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{checkForVServer}
\alias{checkForVServer}
\title{Check for Video Server binary}
\usage{
checkForVServer(jarloc = NULL, update = FALSE, method = "internal",
  extra = getOption("download.file.extra"))
}
\arguments{
\item{jarloc}{A directory in which the binary is to be placed. Defaults to the /bin of rDVR package.}

\item{update}{A boolean indicating whether to update the binary if it is present.}

\item{method}{Method to be used for downloading files. Passed to \code{\link{download.file}}}

\item{extra}{character vector of additional command-line arguments for the "wget", "curl" and "lynx" methods. Passed to \code{\link{download.file}}}
}
\description{
\code{checkForVServer}
A utility function to check if the Video Server standalone binary is present.
}
\section{Detail}{
 The Video Server java binary can be found at https://github.com/johndharrison/rDVR. If users would like to create their own please refe to documentation. This convience function downloads the standalone server and places it in the rDVR package directory bin folder by default.
}
\examples{
\dontrun{
checkForVServer()
}
}

