% Generated by roxygen2 (4.0.0): do not edit by hand
\name{startVideoServer}
\alias{startVideoServer}
\title{Start the Video Server.}
\usage{
startVideoServer(jarloc = NULL, savedir = NULL, port = NULL,
  distmode = FALSE, invisible = TRUE)
}
\arguments{
  \item{jarloc}{A directory in which the standalone video
  server binary is located. Defaults to the /bin of rDVR
  package.}

  \item{savedir}{A directory where the user would like
  videos saved to.  If not declared it defaults to the temp
  folder (which varies depending on the OS).}

  \item{port}{The port on which the video server will
  listen. Defaults to 9998.}

  \item{distmode}{You can enable a "distribution" mode for
  the storage of recorded videos that will use the last two
  characters of the filename requested to save video to
  place it in a subfolder. By default, if you want to save
  a video with the name, say, videofile20987 you will end
  up with the file stored at:
  /path/to/dest/folder/videofile20987.mov. If you had
  enabled the distribution mode with distmode = TRUE the
  video would be stored at:
  /path/to/dest/folder/87/videofile20987.mov}

  \item{invisible}{Windows Only Show the video server in a
  shell. By default it is invisible with setting TRUE.}
}
\description{
\code{startVideoServer}
A utility function to start the standalone video server.
}
\section{Detail}{
  By default the binary is assumed to be in the rDVR
  package /bin directory.
}
\examples{
\dontrun{
startVideoServer()
}
}

