% Generated by roxygen2 (4.0.0): do not edit by hand
\name{checkForVServer}
\alias{checkForVServer}
\title{Check for Video Server binary}
\usage{
checkForVServer(jarloc = NULL, update = FALSE)
}
\arguments{
  \item{jarloc}{A directory in which the binary is to be
  placed. Defaults to the /bin of rDVR package.}

  \item{update}{A boolean indicating whether to update the
  binary if it is present.}
}
\description{
\code{checkForVServer}
A utility function to check if the Video Server standalone binary is present.
}
\section{Detail}{
  The Video Server java binary can be found at
  https://github.com/johndharrison/rDVR. If users would
  like to create their own please refe to documentation.
  This convience function downloads the standalone server
  and places it in the rDVR package directory bin folder by
  default.
}
\examples{
\dontrun{
checkForVServer()
}
}

